# Improve constant-time hash equality