This PR improves the handling of constant-time equality for hashes.

It replaces [`constant_time_eq`](https://crates.io/crates/constant_time_eq) with the more standard [`subtle`](https://crates.io/crates/subtle). Doing this allows us to implement the `ConstantTimeEq` trait for `Hash`, which can be useful to satisfy trait bounds. This trait is then used for all `PartialEq` implementations for consistency.

A side benefit is that because `subtle` plays nicely with Miri, there's no need to keep around the fallback constant-time equality functionality.