@oconnor663 2024-08-21
This change looks good in code, but it seems to come with a substantial performance penalty. I'm pasting this new benchmark at the end of `bench.rs`. (I don't know if all the no_inline/black_box shenanigans are really necessary, but it's a habit.)

```rust
#[no_mangle]
#[inline(never)]
pub fn foo(hash1: &blake3::Hash, hash2: &blake3::Hash) -> bool {
    hash1 == hash2
}

#[bench]
fn bench_eq(b: &mut Bencher) {
    let hash1 = blake3::hash(b"foo");
    let hash2 = blake3::hash(b"bar");
    b.iter(|| foo(test::black_box(&hash1), test::black_box(&hash2)));
}
```

On master I see:

```
$ cargo +nightly bench bench_eq
test bench_eq                          ... bench:           1.37 ns/iter (+/- 0.06)
```

On this branch I see:

```
$ cargo +nightly bench bench_eq
test bench_eq                          ... bench:          31.80 ns/iter (+/- 0.71)
```

31 nanoseconds might not sound like a lot, but for comparison that's almost as long as it takes to _hash_ 64 bytes on my laptop (with AVX-512 support):

```
$ cargo +nightly bench bench_atonce_0001_block
test bench_atonce_0001_block           ... bench:          45.05 ns/iter (+/- 0.33) = 1422 MB/s
```

If we look at `cargo +nightly asm --bench bench foo` and (on your branch) `cargo asm --lib ct_eq` we can see the difference. `constant_time_eq` is able to use SIMD (only SSE2, so there's still room for improvement if we want to get fancy), but `subtle` is doing approximately this for each pair of _bytes_:

```
        movzx eax, byte ptr [r14 + 1]
        xor edi, edi
        cmp al, byte ptr [rbx + 1]
        sete dil
        call subtle::black_box
        mov r15d, eax
        and r15b, bpl
```

Have you looked at these sorts of details before? Do you know if there's anything we can do differently in how we're calling `subtle`, or would this need to be fixed upstream?

> A side benefit is that because subtle plays nicely with Miri, there's no need to keep around the fallback constant-time equality functionality.

This does sound nice. If we don't end up landing this change, I might try to contribute something similar to `constant_time_eq`.