@AaronFeickert 2024-08-21
The optimization barrier in `subtle` used to be inline assembly, but it was changed a few years back (and has seen some iteration). I've seen discussion about returning to that behavior, with initial concerns about stability and portability. Unfortunately, the use of `subtle` would currently require the performance hit. That being said, it is pretty standard at this point in the Rust cryptography ecosystem.

The core library does include its own [optimization barrier](https://doc.rust-lang.org/stable/core/hint/fn.black_box.html) that was previously available in `subtle`. However, a warning was added not to use it for cryptographic purposes, so it's no longer made available.

One option that I mulled over was whether it's reasonably safe enough to use variable-time equality testing for `PartialEq`, and reserve constant-time equality testing for `ConstantTimeEq`. This has the downside of being breaking and easier to misuse, but it does allow implementers to decide if their use case requires it for improved flexibility.