@AaronFeickert 2024-08-21
Anyway, I'm happy to close this if the performance hit isn't worth the library change and Miri cleanup.