@oconnor663 2024-08-22
Let's see what Cesar thinks of that PR I just put up. Also if some callers need trait compatibility with `subtle::ConstantTimeEq`, another option could be to make `subtle` an optional feature and implement it as a wrapper around the current implementation.