@AaronFeickert 2024-08-22
Good point overall; the only proper approach would be to nix it entirely, so hopefully your PR gets accepted!

I do think that the overall risk is still greater (at least marginally) with the current workaround, especially given that the function is named [`constant_time_eq_miri`](https://github.com/BLAKE3-team/BLAKE3/blob/479eef82d756221abe9f93077b91ce8cf763c32a/src/lib.rs#L322) and there's no reason to think it couldn't or wouldn't ever be variable time. Unless it's too much of a nit, I'll go ahead and open a PR to remove it now in favor of standard variable-time `PartialEq` on the underlying byte array; a scary comment will accompany it!

Separately, good idea that `subtle` could be used as an optional feature, to get its useful traits for callers who need or want them. If that were the case, I think it would be proper to have its constant-time trait implementations use `subtle` directly and _not_ wrap `constant_time_eq`, in order to ensure the caller is getting what it says on the box. This would have the extra advantage of allowing the existing `PartialEq` functionality to retain its speedy behavior. The extra cost would be the extra dependency, but at least it would be optional.