@cesarb 2024-08-27
> Let's see what Cesar thinks of that PR I just put up.

Sorry for the delay. Looks fine to me. Merged and published `constant_time_eq` 0.3.1 with only that change.