@AaronFeickert 2024-08-28
@oconnor663: After the ongoing discussion, this has been updated to retain the existing (speedy) functionality from `constant_time_eq`, but add a `ConstantTimeEq` implementation for `Hash` via a new optional `subtle` trait.

I think this is a useful addition, and also opens the door to adding other useful traits in the future if desired.