@AaronFeickert 2025-01-04
@BurningEnlightenment: Thanks! This PR has been in the queue for the while; would appreciate any insight into whether or not it is useful enough to keep around.