@oconnor663 2025-04-01
Sorry for letting this sit. Two questions: Even though we're implementing the trait from `subtle`, can we use `constant_time_eq` to do the comparison internally? And should we fold this into the existing `traits-preview` feature, which enables trait implementations from the `digest` crate? I know that `digest` is a RustCrypto project and `subtle` is not, but on the other hand `digest` does take a dependency on `subtle`.