@AaronFeickert 2025-04-01
We certainly could use the existing `constant_time_eq` functionality, but that doesn't really sit well with me. There's no strict requirement that `ConstantTimeEq` be implemented in any particular way (aside from attempting to execute in constant time), but I wager users of the library would expect that a `subtle` feature implementing `subtle` traits would use `subtle` functionality where available, especially since their use of those traits likely implies their trust in `subtle` functionality.

I also don't think baking it in with the existing `traits-preview` feature is the right move, since `subtle` is stable and independent from the digest traits. Keeping it separate also has the advantage of allowing for a (slightly) smaller dependency footprint.

But I am open to discussion on these points!