@oconnor663 2025-04-20
For more context for me, what's an example of a caller that needs an implementation of the `Subtle` trait specifically (i.e. in a generic context), as opposed to just "we promise equality is constant time" in the docs?